kidneyimmune

kidneyimmune - A small set of demo data from Stewart et al. 2019 Science. See https://www.kidneycellatlas.org/

means - Dataframe of CellPhoneDB output means.txt file

pvals - Dataframe of CellPhoneDB output pvalues.txt file

decon - Dataframe of CellPhoneDB output deconvoluted.txt file

sig_means - Dataframe of CellPhoneDB output significant_means.txt file

means2 - Dataframe of CellPhoneDB output means.txt file

pvals2 - Dataframe of CellPhoneDB output pvalues.txt file

decon2 - Dataframe of CellPhoneDB output deconvoluted.txt file

sig_means2 - Dataframe of CellPhoneDB output significant_means.txt file

covid_cpdb_meta - Example dataframe to use for cpdb_meta option in compare_cpdb

covid_sample_metadata - Example dataframe to use for sample_metadata option in compare_cpdb

cpdb_output_v5 - Dataframe of CellPhoneDB output means.txt file

sce_v5 - A small dummy singlecelldata for cellphonedb v5

relevant_interactions_v5 - Dataframe of CellPhoneDB output relevant_interactions.txt file

decon_v5 - Dataframe of CellPhoneDB output deconvoluted.txt file

cellsign_v5 - Dataframe of CellPhoneDB output CellSign.txt file

interaction_scores_v5 - Dataframe of CellPhoneDB output interaction_scores.txt file

data(kidneyimmune)

data(cpdb_output)

pvals

decon

sig_means

data(cpdb_output)

pvals2

decon2

sig_means2

covid_cpdb_meta

covid_sample_metadata

data(cpdb_output_v5)

sce_v5

relevant_interactions_v5

decon_v5

cellsign_v5

interaction_scores_v5

Format

A SingleCellExperiment object with the following slots filled

assays

counts - Raw expression data

logcounts - Normalized expression data

colData

Cell level metadata

Data frames containing outputs after CellPhoneDB analysis

means

data frame containing mean expression values for each interacting pair

pvals

data frame containing p values for each interacting pair

An object of class data.frame with 837 rows and 11892 columns.

An object of class data.frame with 1965 rows and 115 columns.

An object of class data.frame with 837 rows and 11893 columns.

Data frames containing outputs after CellPhoneDB analysis

means2

data frame containing mean expression values for each interacting pair

pvals2

data frame containing p values for each interacting pair

An object of class data.frame with 837 rows and 155 columns.

An object of class data.frame with 1965 rows and 18 columns.

An object of class data.frame with 837 rows and 156 columns.

An object of class data.frame with 12 rows and 3 columns.

An object of class data.frame with 12 rows and 2 columns.

data after CellPhoneDB v5 analysis

An object of class SingleCellExperiment with 30800 rows and 3312 columns.

An object of class data.frame with 419 rows and 85 columns.

An object of class data.frame with 6825 rows and 49 columns.

An object of class data.frame with 9 rows and 85 columns.

An object of class data.frame with 2440 rows and 94 columns.

Source

https://www.kidneycellatlas.org/

Examples

data(kidneyimmune)
data(cpdb_output)
data(cpdb_output)
data(cpdb_output_v5)